Update product microservice docs and API mapping

+++ product_management/src/main/java/com/example1/product_management/controller/ProductsController.java
@@ -1,11 +1,10 @@
 package com.example1.product_management.controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 import com.example1.product_management.dto.ProductResponseDTO;
 import com.example1.product_management.dto.ProductRequestDTO;
 import com.example1.product_management.service.ProductService;
-
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +12,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
+import com.example1.product_management.exceptions.NotFoundException;
+import com.example1.product_management.exceptions.BussinesException;
 import java.util.List;
 
 
@@ -34,8 +37,11 @@ public class ProductsController {
      * @return Lista de productos
      */
     @GetMapping()
-    public List<ProductResponseDTO> Products() {
-        return productService.getProducts();
+    public ResponseEntity<List<ProductResponseDTO>> getProducts() {
+        if (productService.getProducts().isEmpty()) {
+            throw new NotFoundException("No products found", "P-404", HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(productService.getProducts(), HttpStatus.OK);
     }
 
     /**
@@ -44,8 +50,12 @@ public class ProductsController {
      * @return Producto encontrado
      */
     @GetMapping("/{id}")
-    public ProductResponseDTO getProductbyId(@PathVariable Long id) {
-        return productService.getProductById(id);
+    public ResponseEntity<ProductResponseDTO> getProductbyId(@PathVariable Long id) {
+        ProductResponseDTO product = productService.getProductById(id);
+        if (product == null) {
+            throw new NotFoundException("Product not found", "P-404", HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(product, HttpStatus.OK);
     }
 
     /**
@@ -54,8 +64,11 @@ public class ProductsController {
      * @return Producto creado
      */
     @PostMapping()
-    public ProductResponseDTO createProduct(@RequestBody ProductRequestDTO product) {
-        return productService.createProduct(product);
+    public ResponseEntity<ProductResponseDTO> createProduct(@RequestBody ProductRequestDTO product) {
+        if (product.getName() == "" || product.getPrice() == 0) {
+            throw new BussinesException("Name and price are required", "P-400", HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(productService.createProduct(product), HttpStatus.CREATED);
     }
 
     /**
@@ -65,8 +78,11 @@ public class ProductsController {
      * @return Producto actualizado
      */
     @PutMapping("/{id}")
-    public ProductResponseDTO updateProduct(@PathVariable Long id, @RequestBody ProductRequestDTO product) {
-        return productService.updateProduct(id, product);
+    public ResponseEntity<ProductResponseDTO> updateProduct(@PathVariable Long id, @RequestBody ProductRequestDTO product) {
+        if (productService.getProductById(id) == null) {
+            throw new NotFoundException("Product not found", "P-404", HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(productService.updateProduct(id, product), HttpStatus.OK);
     }
 
     /**
@@ -74,8 +90,26 @@ public class ProductsController {
      * @param id
      */
     @DeleteMapping("/{id}")
-    public void deleteProduct(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+        if (productService.getProductById(id) == null) {
+            throw new NotFoundException("Product not found", "P-404", HttpStatus.NOT_FOUND);
+        }
         productService.deleteProduct(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+    /**
+     * Obtener un producto por nombre
+     * @param name
+     * @return Producto encontrado
+     */
+
+    @GetMapping(params = "name")
+    public ResponseEntity<ProductResponseDTO> getProductByName(@RequestParam("name") String name) {
+        if (productService.getProductByName(name) == null) {
+            throw new NotFoundException("Product not found", "P-404", HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(productService.getProductByName(name), HttpStatus.OK);
     }
 
 }
